feat(store): add query_ovs_db method to store trait
The `query_data` method makes no sense when using databases since it uses
different types of lists to query the required data, when that's a single query
in a database; thereby we are adding the `query_ovs_db` method which explicitly
queries the required data from the database.

We are also adding a new error type `StoreError::MethodNotAvailable` that will
be yielded when we call `query_data` from a database store implementation, or
when we call the new `query_ovs_db` method from a directory store
implementation.

Signed-off-by: Irene Diez <idiez@redhat.com>